move elements for formfooter to componenet
