+++ frontend/src/components/posts/PostForm.jsx
@@ -6,6 +6,7 @@ import ResponsiveTextArea from '../ResponsiveTextArea';
 
 // Project Files
 import PostsApi from "../../api/PostsApi";
+import FormFooter from "../FormFooter";
 
 export default function PostForm({ onSubmit, onSubmitMedia }) {
   // Local State
@@ -40,13 +41,6 @@ export default function PostForm({ onSubmit, onSubmitMedia }) {
     setIsFilePicked(true);
   };
 
-  function getNiceName(requiredLength) {
-    if (requiredLength >= contentFile.name.length) {
-      return contentFile.name;
-    } else {
-      return contentFile.name.substring(0, requiredLength) + "..";
-    }
-  }
 
   return (
     <div className="form-container">
@@ -58,49 +52,14 @@ export default function PostForm({ onSubmit, onSubmitMedia }) {
             onFormContentChange={onFormContentChange}
             maxLength="1000"
           />
-          <div className="media-name-character-limit-display">
-            <div className={`${isFilePicked ? "file-selected" : "no-file"}`}>
-              <span className="media-name-full">
-                📎 {isFilePicked ? getNiceName(35) : "nothing selected"}
-              </span>
-              <span className="media-name-mobile">
-                📎 {isFilePicked ? getNiceName(16) : "nothing selected"}
-              </span>
-            </div>
-            <p className="length">{length == null ? 0 : length}/255</p>
-          </div>
         </div>
-
-        <div className="form-footer">
-          <button className="button-post" onClick={handleSubmit}>
-            Post
-          </button>
-          <div className="form-footer-extras">
-            <input
-              hidden
-              id="pick-media-image"
-              className="pick-media"
-              type="file"
-              onChange={setFile}
-              accept="image/*"
-            />
-            <label for="pick-media-image" className="proxy-button pick-image">
-              <span className="tooltiptext">pick image</span>
-            </label>
-            <input
-              hidden
-              id="pick-media-video"
-              className="pick-media"
-              type="file"
-              onChange={setFile}
-              accept="video/*"
-            />
-            <label for="pick-media-video" className="proxy-button pick-video">
-              <div></div>
-              <span className="tooltiptext">pick video</span>
-            </label>
-          </div>
-        </div>
+        <FormFooter
+          isFilePicked={isFilePicked}
+          contentFile={contentFile}
+          length={length}
+          setFile={setFile}
+          handleSubmit={handleSubmit}
+        />
       </form>
     </div>
   );
